New low-level topology setup configuration
There is a new method to perform topology setup. The new method opens a
new channel in the consumer collection and expose the channel setup
function to the clients to enable them to use methods like assertQueue
to describe the full topology. We decided to expose this functionality
to keep the new producer and consumer APIs clean. The new topology
method is experimental and is very likely that we replace it for a
simpler API in the future.