Fix a bug in connection management
In the old, deprecated version of the publisher, we were opening a new
channel everytime a message was published to the broker, leaking
channels and eventually slowing down the server. We discovered this
issue when Heimdall started to use the 0.7x branch of Elementals.
Although the bug is not present in the new publisher API, in order to
ease the transition it is necessary to fix the bug in the old method.