Improvements to gather more metrics
The first change is to initialize the errors metric to zero. This is
done to avoid "holes" or missing data in the errors time series, to
improve reports.

The second change is to add timestamps when messages are published. We
are going to use that information later in order to track how long it
takes to process messages.