Improve RabbitMQ metrics
This commit changes the way we gather Prometheus metrics for RabbitMQ.
We are now taking advantage of the timestamp published inside messages
to track how long does it take to process them (since they are published
until they are consumed and acked). Additionally, we had two different
timeseries for incoming and two differente timeseries for outgoing
messages, tracking successfull and failed operations as different time
series. With this commit we unify them into a single time series and we
use a label to indicate if the operation succeeded or failed. This
should simplify computations like success rate.