Improve RabbitMQ metrics
This commit restores the incoming and failed message counters, to
initialize them to zero when the application starts and avoid missing
metrics. Also, the histogram that tracks message lifetimes was splitted
in two: one that tracks how long messages remain queued and another one
that tracks how long does it takes consumers to process the messages.